requirements.txt
discord.py==2.3.2
aiohttp==3.9.3
SQLAlchemy[asyncio]==2.0.27
aiosqlite==0.20.0
Pillow==10.2.0
wavelink==3.2.0
spotipy==2.23.0
google-api-python-client==2.118.0
tweepy==4.14.0
praw==7.7.1
asyncpraw==7.7.1
google-cloud-language==2.13.0
python-dotenv==1.0.1
aiofiles==23.2.1
pytz==2024.1
humanfriendly==10.0
dateparser==1.2.0
numpy==1.26.4
requests==2.31.0
pydub==0.25.1
PyNaCl==1.5.0
instaloader==4.10.3
yt-dlp==2024.3.10
